net/gnrc/ppp: adapted old ppp to netif

+++ sys/include/net/gnrc/ppp/opt.h
@@ -107,12 +107,12 @@ static inline void ppp_opt_set_payload(gnrc_ppp_option_t *opt, void *data, size_
  *
  * @return pointer to next option
  */
-static inline gnrc_ppp_option_t *ppp_opt_get_next(gnrc_ppp_option_t *curr_opt, gnrc_ppp_option_t *head, size_t opt_size)
+static inline gnrc_ppp_option_t *ppp_opt_get_next(gnrc_ppp_option_t *curr_opt, gnrc_ppp_option_t *head, ssize_t opt_size)
 {
     gnrc_ppp_option_t *ret = NULL;
 
     ret = (gnrc_ppp_option_t *)(((uint8_t *) curr_opt) + ppp_opt_get_length(curr_opt));
-    if (ret - head >= opt_size) {
+    if (((int)ret - (int)head) >= opt_size) {
         return NULL;
     }
     return ret;
